backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
alembic
python-dotenv

# NEW - File handling and storage
minio==7.2.0
python-multipart==0.0.6
python-magic==0.4.27
aiofiles==23.2.1

# NEW - Text extraction
PyPDF2==3.0.1
pypdfium2==4.26.0
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0
openpyxl==3.1.2

# NEW - Async task processing
celery==5.3.4
redis==5.0.1
flower==2.0.1

openai==1.3.0
tiktoken==0.5.1